services/device-verification-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
redis>=5.2.0
python-dotenv>=1.0.0

# OpenTelemetry
opentelemetry-api>=1.28.2
opentelemetry-sdk>=1.28.2
opentelemetry-instrumentation-fastapi>=0.49b2
opentelemetry-instrumentation-httpx>=0.49b2
opentelemetry-exporter-otlp-proto-http>=1.28.2

